Clean up systrace a bit.
Extract the various event definitions into a table instead of a
switch statement.

Add a --link-assets option to systrace.py to generate a trace.html
file that has the CSS and JS assets linked in rather than embedded,
which is useful for debugging purposes.

Fixed a bug in the tests for lineRE.

Bug: 6488845
Change-Id: I90634f7144fa0a050b911bb9854e5d5662dd6330