systrace: rebase against upstream trace-viewer
This changes replaces the use of the Chromium trace viewer code with that of
the trace-viewer project (essentially the Chromium code factored out into its
own project).  It updates the update.py script to pull the source from there
and adds an argument to systrace.py to use linked assets from a different
directory to facilitate development of trace-viewer code using systrace to
test.

Change-Id: I5c5a4ce336ce82cd07f1f42e97d4b4cb918ec17a