Fix thread names when absent or truncated.
When a large number of threads are active, the kernel may
start to report <...> as the name of a thread in the trace
buffer which is rather unhelpful.

This change uses information from 'ps -t' to fix thread names
when they are absent or truncated.  Use the '--no-fix-threads'
option to disable this feature.

Change-Id: I6566e557eeac33311ff5d4c6e95ff010cb83a0d6