Added ability to import traces from multiple tracers.
The current design of systrace can only process traces from one tracer, atrace.
This change makes it possible to import multiple tracers that can be stored in
the trace.html output. Each trace is now stored as a script tag with class
"trace-data".

This patch also simplifies storing traces in the trace.html output. Instead of
storing traces in a javascript variable, they are now simply stored as scripts
with type "application/text".

Change-Id: I39efd7e1a512fb37f84ff0c7ef6105e13bf1491b