Add comments and reformatting to get rid of lint warnings.
Added a few comments and cleaned up most lint warnings.

I was not able to fix lint warnings about the interpreter and
copyright blocks (pylint comments to disable those warnings also did not work).

I left in a few warnings in code that is being substantially changed as
part of another CL.

Change-Id: I365d8caeb264521c45b5995a24951b0da8eb4c29