Added --no-compress option.
Added an option to download the trace data in plain text,
instead of in zipped form (--no-compress).  This may be
necessary if the zipped trace data contains embedded
CR+LF sequences.  This could be a problem in theory, but
in practice this either never happens or is very infrequent.

This is also a handy option to use when debugging this script
and it's communication with adb.

Change-Id: I814e883424234c1e0284459b229447dea0c1a410